--- packages/kinvey-utils/package.json
@@ -1,19 +1,0 @@
-{
-  "private": true,
-  "version": "3.9.3-beta.7",
-  "name": "kinvey-utils",
-  "main": "index.js",
-  "dependencies": {
-    "es6-promise": "4.1.1",
-    "lodash": "4.17.4",
-    "qs": "6.5.1",
-    "rxjs": "5.5.2"
-  },
-  "devDependencies": {
-    "expect": "1.20.2",
-    "mocha": "4.0.1"
-  },
-  "scripts": {
-    "test": "mocha -s 250 test/**/*.spec.js"
-  }
-}
